pyproject.toml
[project]
name = "google-genai-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "ipykernel>=7.2.0",
    "numpy>=2.4.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "sqlite-utils>=3.39",
]
